+++ docs/changelog.md
@@ -89,3 +89,4 @@
 
 
 - Git Push [2025-10-19 08:48:46](.logs/audits/git/git.html#run-20251019-084846) → https://github.com/gunreip/tafeld.git [main] : Initial push
+- Git Push [2025-10-19 10:00:34](.logs/audits/git/git.html#run-20251019-100034) → https://github.com/gunreip/tafeld.git [main] : Upgrade Command ProjectGitPush